+++ src/main/java/com/podio/item/filter/FilterFieldValue.java
@@ -2,7 +2,7 @@ package com.podio.item.filter;
 
 public abstract class FilterFieldValue<VALUES> extends Filter<VALUES> {
 
-    public FilterFieldValue(int fieldId) {
+    public FilterFieldValue(long fieldId) {
         super(String.valueOf(fieldId));
     }
 }
